Pull common UI panel code into a base class.
