+++ traitsui/wx/ui_base.py
@@ -19,92 +19,24 @@
     dialogs.
 """
 
-#-------------------------------------------------------------------------
-#  Imports:
-#-------------------------------------------------------------------------
-
 import wx
 
-from traits.api \
-    import HasStrictTraits, HasPrivateTraits, Instance, List, Event
+from traits.api import HasPrivateTraits, Instance
 
-from traitsui.api \
-    import UI
-
-from traitsui.menu \
-    import Action
-
-from editor \
-    import Editor
-
-#-------------------------------------------------------------------------
-#  Constants:
-#-------------------------------------------------------------------------
-
-# List of all predefined system button names:
-SystemButtons = ['Undo', 'Redo', 'Apply', 'Revert', 'OK', 'Cancel', 'Help']
-
-#-------------------------------------------------------------------------
-#  'RadioGroup' class:
-#-------------------------------------------------------------------------
-
-
-class RadioGroup(HasStrictTraits):
-    """ A group of mutually-exclusive menu or toolbar actions.
-    """
-    # List of menu or tool bar items
-    items = List
-
-    #-------------------------------------------------------------------------
-    #  Handles a menu item in the group being checked:
-    #-------------------------------------------------------------------------
-
-    def menu_checked(self, menu_item):
-        """ Handles a menu item in the group being checked.
-        """
-        for item in self.items:
-            if item is not menu_item:
-                item.control.Check(False)
-                item.item.action.checked = False
-
-    #-------------------------------------------------------------------------
-    #  Handles a tool bar item in the group being checked:
-    #-------------------------------------------------------------------------
-
-    def toolbar_checked(self, toolbar_item):
-        """ Handles a tool bar item in the group being checked.
-        """
-        for item in self.items:
-            if item is not toolbar_item:
-                item.tool_bar.ToggleTool(item.control_id, False)
-                item.item.action.checked = False
-
-#-------------------------------------------------------------------------
-#  'ButtonEditor' class:
-#-------------------------------------------------------------------------
+from traitsui.base_panel import BasePanel as _BasePanel
+from traitsui.menu import Action
+from .editor import Editor
 
 
 class ButtonEditor(Editor):
     """ Editor for buttons.
     """
-    #-------------------------------------------------------------------------
-    #  Trait definitions:
-    #-------------------------------------------------------------------------
-
     # Action associated with the button
     action = Instance(Action)
-
-    #-------------------------------------------------------------------------
-    #  Initializes the object:
-    #-------------------------------------------------------------------------
 
     def __init__(self, **traits):
         # XXX Why does this need to be an Editor subclass? -- CJW
         HasPrivateTraits.__init__(self, **traits)
-
-    #-------------------------------------------------------------------------
-    #  Handles the associated button being clicked:
-    #-------------------------------------------------------------------------
 
     def perform(self, event):
         """ Handles the associated button being clicked.
@@ -113,17 +45,14 @@ class ButtonEditor(Editor):
         self.ui.do_undoable(handler.perform, self.ui.info, self.action, event)
 
 
-#-------------------------------------------------------------------------
-#  'BaseDialog' class:
-#-------------------------------------------------------------------------
-
-
-class BaseDialog(object):
+class BaseDialog(_BasePanel):
     """ Base class for Traits UI dialog boxes.
     """
-    #-------------------------------------------------------------------------
-    #  Sets the frame's icon:
-    #-------------------------------------------------------------------------
+
+    def default_icon(self):
+        """ Return a default icon for a TraitsUI dialog. """
+        from pyface.image_resource import ImageResource
+        return ImageResource('frame.ico')
 
     def set_icon(self, icon=None):
         """ Sets the frame's icon.
@@ -131,12 +60,8 @@ class BaseDialog(object):
         from pyface.image_resource import ImageResource
 
         if not isinstance(icon, ImageResource):
-            icon = ImageResource('frame.ico')
+            icon = self.default_icon()
         self.control.SetIcon(icon.create_icon())
-
-    #-------------------------------------------------------------------------
-    #  Adds a status bar to the dialog:
-    #-------------------------------------------------------------------------
 
     def add_statusbar(self):
         """ Adds a status bar to the dialog.
@@ -178,10 +103,6 @@ class BaseDialog(object):
 
         return set_status_text
 
-    #-------------------------------------------------------------------------
-    #  Adds a menu bar to the dialog:
-    #-------------------------------------------------------------------------
-
     def add_menubar(self):
         """ Adds a menu bar to the dialog.
         """
@@ -192,10 +113,6 @@ class BaseDialog(object):
                 menubar.create_menu_bar(self.control, self))
             self._last_group = self._last_parent = None
 
-    #-------------------------------------------------------------------------
-    #  Adds a tool bar to the dialog:
-    #-------------------------------------------------------------------------
-
     def add_toolbar(self):
         """ Adds a toolbar to the dialog.
         """
@@ -205,117 +122,6 @@ class BaseDialog(object):
             self.control.SetToolBar(
                 toolbar.create_tool_bar(self.control, self))
             self._last_group = self._last_parent = None
-
-    #-------------------------------------------------------------------------
-    #  Adds a menu item to the menu bar being constructed:
-    #-------------------------------------------------------------------------
-
-    def add_to_menu(self, menu_item):
-        """ Adds a menu item to the menu bar being constructed.
-        """
-        item = menu_item.item
-        action = item.action
-
-        if action.id != '':
-            self.ui.info.bind(action.id, menu_item)
-
-        if action.style == 'radio':
-            if ((self._last_group is None) or
-                    (self._last_parent is not item.parent)):
-                self._last_group = RadioGroup()
-                self._last_parent = item.parent
-            self._last_group.items.append(menu_item)
-            menu_item.group = self._last_group
-
-        if action.enabled_when != '':
-            self.ui.add_enabled(action.enabled_when, menu_item)
-
-        if action.checked_when != '':
-            self.ui.add_checked(action.checked_when, menu_item)
-
-    #-------------------------------------------------------------------------
-    #  Adds a tool bar item to the tool bar being constructed:
-    #-------------------------------------------------------------------------
-
-    def add_to_toolbar(self, toolbar_item):
-        """ Adds a toolbar item to the toolbar being constructed.
-        """
-        self.add_to_menu(toolbar_item)
-
-    #-------------------------------------------------------------------------
-    #  Returns whether the menu action should be defined in the user interface:
-    #-------------------------------------------------------------------------
-
-    def can_add_to_menu(self, action, action_event=None):
-        """ Returns whether the action should be defined in the user interface.
-        """
-        if action.defined_when == '':
-            return True
-
-        return self.ui.eval_when(action.defined_when)
-
-    #-------------------------------------------------------------------------
-    #  Returns whether the toolbar action should be defined in the user
-    #  interface:
-    #-------------------------------------------------------------------------
-
-    def can_add_to_toolbar(self, action):
-        """ Returns whether the toolbar action should be defined in the user
-            interface.
-        """
-        return self.can_add_to_menu(action)
-
-    #-------------------------------------------------------------------------
-    #  Performs the action described by a specified Action object:
-    #-------------------------------------------------------------------------
-
-    def perform(self, action, event):
-        """ Performs the action described by a specified Action object.
-        """
-        handler = self.ui.handler
-        self.ui.do_undoable(handler.perform, self.ui.info, action, event)
-
-    #-------------------------------------------------------------------------
-    #  Check to see if a specified 'system' button is in the buttons list, and
-    # add it if it is not:
-    #-------------------------------------------------------------------------
-
-    def check_button(self, buttons, action):
-        """ Adds *action* to the system buttons list for this dialog, if it is
-        not already in the list.
-        """
-        name = action.name
-        for button in buttons:
-            if self.is_button(button, name):
-                return
-        buttons.append(action)
-
-    #-------------------------------------------------------------------------
-    #  Check to see if a specified Action button is a 'system' button:
-    #-------------------------------------------------------------------------
-
-    def is_button(self, action, name):
-        """ Returns whether a specified action button is a system button.
-        """
-        if isinstance(action, basestring):
-            return (action == name)
-        return (action.name == name)
-
-    #-------------------------------------------------------------------------
-    #  Coerces a string to an Action if necessary:
-    #-------------------------------------------------------------------------
-
-    def coerce_button(self, action):
-        """ Coerces a string to an Action if necessary.
-        """
-        if isinstance(action, basestring):
-            return Action(name=action,
-                          action='?'[(not action in SystemButtons):])
-        return action
-
-    #-------------------------------------------------------------------------
-    #  Creates a user specified button:
-    #-------------------------------------------------------------------------
 
     def add_button(self, action, sizer, method=None, enabled=True,
                    name=None, default=False):
